requirements.txt
# Required packages for transcript cleaning
openai>=1.0.0
python-dotenv>=1.0.0
